changed how the users api endpoint was set. added the rest of users endpoints

+++ api/user/delete_user.php
@@ -3,11 +3,17 @@ declare(strict_types=1);
 
 header('Content-Type: application/json');
 
-echo $id;
+use App\USers\UserService;
 
-use App\Service\UserService;
-
+$id = (int) $id;
 $userService = new UserService();
+
+if (!$userService->exists($id)) {
+	http_response_code(404);
+	header('Content-Type: application/json');
+	echo json_encode(['error' => 'User not found']);
+	exit();
+}
 
 $success = $userService->delete($id);
